+++ GooglePay/src/main/java/com/braintreepayments/api/GooglePayPaymentAuthRequestCallback.java
@@ -1,13 +1,10 @@
 package com.braintreepayments.api;
 
-import androidx.annotation.Nullable;
-
 /**
  * Callback to handle result from
  * {@link GooglePayClient#createPaymentAuthRequest(GooglePayRequest, GooglePayPaymentAuthRequestCallback)}
  */
 public interface GooglePayPaymentAuthRequestCallback {
 
-    void onResult(@Nullable GooglePayPaymentAuthRequest paymentAuthRequest,
-                  @Nullable Exception error);
+    void onGooglePayPaymentAuthRequest(GooglePayPaymentAuthRequest paymentAuthRequest);
 }