+++ GooglePay/src/main/java/com/braintreepayments/api/GooglePayPaymentAuthRequestParams.java
@@ -6,14 +6,14 @@ import com.google.android.gms.wallet.PaymentDataRequest;
 
 /**
  * Used to request Google Pay payment authorization via
- * {@link GooglePayLauncher#launch(GooglePayPaymentAuthRequest)}
+ * {@link GooglePayLauncher#launch(GooglePayPaymentAuthRequestParams)}
  */
-public class GooglePayPaymentAuthRequest {
+public class GooglePayPaymentAuthRequestParams {
 
     private final int googlePayEnvironment;
     private final PaymentDataRequest paymentDataRequest;
 
-    GooglePayPaymentAuthRequest(int googlePayEnvironment, @NonNull PaymentDataRequest paymentDataRequest) {
+    GooglePayPaymentAuthRequestParams(int googlePayEnvironment, @NonNull PaymentDataRequest paymentDataRequest) {
         this.googlePayEnvironment = googlePayEnvironment;
         this.paymentDataRequest = paymentDataRequest;
     }